app/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg[binary]
python-dotenv
redis
url-normalize